scripts/gruntz/audit/jcc_sieve.py - reads report.json for scores, disassembles
the build/objdiff/{base,target} object pairs (one llvm-objdump per OBJECT, not
per symbol) and diffs the ordered (offset, mnemonic, target) branch list.
~6s for the whole tree.

    jcc sieve: 173 hits over 1291 sub-100% functions | SIGNEDNESS 2 TOPOLOGY 7
               OTHER 27 POLARITY 137 (101 pure je/jne)
      not this signal: 327 differ in branch COUNT | 27 >4 flips | 88 branchless
                       | 674 identical branch sequence
      uncomparable: 2 undecodable (jump-table data in .text)

Beyond the brief it added TOPOLOGY (every mnemonic agrees but a branch lands on
a different block - the SetDimensions shape, the one --diff hides hardest, with
targets named symbolically so uniform displacement shifts compare equal) and per
-row ret counts so the shrink-wrap test can be read off the worklist.

SIX BUGS THAT WERE NOT CODEGEN PREFERENCE:

1. ProjectWallQuad wrote 32 bytes BEFORE the array. Both vertex passes rewrite
   each record in place; we stored to v[-8]/v[-7] - out of bounds on iteration 0
   - and ran the translate pass over records 3..6 instead of 0..3.
2. CRandomAmbientSound::Update had BOTH proximity-box tests inverted: a random
   ambient sound was silenced inside its own zone and played everywhere outside.
   The inverted version scores 96.56 and the correct one 95.04 - correctness
   kept, as with BuildPowerupIconKeys earlier today.
3. CTileTriggerLogic::Classify returned +1 where retail's shared block is
   `or eax,-1`: it reported "still active" on an expired duty cycle. A prior note
   asserted the opposite citing "the full retail decode"; the branch TARGET
   disproves it, which is exactly what --diff masks.
4. SoundStream::ParseWave's fmt cap was signed - a chunk size above 0x7fffffff
   counted as "small" and was read into an 18-byte WaveFormatX.
5. PruneOrphans compared a slot against itself (the Lookup miss arm re-zeroed
   `found`), where retail short-circuits.
6. The inlined MFC CArray::SetSize grow clamp is >=, not > - five open-coded
   copies.

g_frameDelta is u32: every retail comparison is jb/jbe/jae, the /9 divide is
mul+shr not imul+sar, and every (double) cast lowers as a zero-extended fild
qword. It was declared i32 in FIVE headers and papered over with static_cast<u32>
at 16 sites - four duplicate externs deleted, 16 casts gone.

SIX "unsteerable walls" were source shapes after all - CFaderSine::ApplyInit
(one ||, +15.5), AutoTuneCmdDelay (+6.2), CRezImage::FlipVertical (+5.8),
CParseSource::Read, CTimeBomb::LoadAttributes, and CMenuPage::FocusNext/Prev
which went 97.18 -> 100.00 EXACT.

Also caught a bug in its own first cut: `^j(?!mp\b)` lets `jmpl` through, which
put unconditional jumps in the sequence. Whitelist now, and an unrecognised j*
raises rather than being silently dropped.

exact 2980 -> 2982 (70.0%), 77.74% fuzzy. `gruntz build --full` exit 0.

+++ src/Gruntz/FontConfig.cpp
@@ -353,7 +353,7 @@ i32 CFontConfig::RenderInputText(HDC hdc, i32 maxWidth, RECT* rect) {
             }
         }
         i32 t;
-        if (static_cast<u32>(g_frameDelta) < static_cast<u32>(g_caretBlinkMs)) {
+        if (g_frameDelta < static_cast<u32>(g_caretBlinkMs)) {
             t = g_caretBlinkMs - g_frameDelta;
         } else {
             t = 0;
